# openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/google:
    get:
      operationId: AuthController_googleLogin
      parameters: []
      responses:
        '200':
          description: ''
        '302':
          description: Redireciona para o Google OAuth
      summary: Inicia o fluxo OAuth Google para membros
      tags: &ref_0
        - auth
  /auth/google/callback:
    get:
      operationId: AuthController_googleCallback
      parameters:
        - name: invite
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Login bem-sucedido, retorna tokens de acesso
        '401':
          description: Autenticação falhou
      summary: Callback do OAuth Google — retorna JWT e dados do usuário
      tags: *ref_0
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Novos tokens gerados com sucesso
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Refresh token inválido ou revogado
      security:
        - bearer: []
      summary: Renova o access token usando o refresh token
      tags: *ref_0
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Sessão encerrada com sucesso
        '201':
          description: ''
      security:
        - bearer: []
      summary: Revoga o refresh token e encerra a sessão
      tags: *ref_0
...
  /admin/invites:
    post:
      operationId: InvitesController_createInvite
      parameters: []
      responses:
        '201':
          description: Convite criado com sucesso
      security:
        - bearer: []
      summary: Cria um novo convite e retorna o link com token
      tags:
        - Invites
        - admin/invites
    get:
      operationId: InvitesController_listInvites
      parameters: []
      responses:
        '200':
          description: Lista de convites
      security:
        - bearer: []
      summary: Lista todos os convites gerados
      tags:
        - Invites
        - admin/invites
  /admin/invites/{id}:
    delete:
      operationId: InvitesController_invalidateInvite
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Convite invalidado
        '404':
          description: Convite não encontrado
      security:
        - bearer: []
      summary: Invalida um convite existente
      tags:
        - Invites
        - admin/invites
  /invites/{token}:
    get:
      operationId: InvitesController_validateToken
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Status do convite
        '404':
          description: Convite não encontrado ou inválido
      summary: Valida um token de convite (público)
      tags:
        - Invites
        - invites
  /admin/auth/google:
    get:
      operationId: AdminAuthController_googleLogin
      parameters: []
      responses:
        '200':
          description: ''
        '302':
          description: Redireciona para o Google OAuth (admin)
      summary: Inicia o fluxo OAuth Google para administradores
      tags: &ref_1
        - admin/auth
  /admin/auth/google/callback:
    get:
      operationId: AdminAuthController_googleCallback
      parameters: []
      responses:
        '200':
          description: Login admin bem-sucedido
        '403':
          description: Usuário não possui role admin
      summary: Callback OAuth Google (admin) — valida role e retorna JWT
      tags: *ref_1
  /admin/auth/refresh:
    post:
      operationId: AdminAuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Novos tokens gerados
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Refresh token inválido
      security:
        - bearer: []
      summary: Renova o access token (admin)
      tags: *ref_1
  /admin/auth/logout:
    post:
      operationId: AdminAuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Sessão encerrada
        '201':
          description: ''
      security:
        - bearer: []
      summary: Encerra a sessão do admin
      tags: *ref_1
...
  /admin/rewards:
    get:
      operationId: AdminRewardsController_list
      parameters: []
      responses:
        '200':
          description: Lista de recompensas
      security: &ref_4
        - bearer: []
      summary: Lista todas as recompensas (admin)
      tags: &ref_5
        - admin/rewards
    post:
      operationId: AdminRewardsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRewardDto'
      responses:
        '201':
          description: Recompensa criada
      security: *ref_4
      summary: Cria uma nova recompensa
      tags: *ref_5
  /admin/rewards/{id}:
    get:
      operationId: AdminRewardsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados da recompensa
        '404':
          description: Recompensa não encontrada
      security: *ref_4
      summary: Retorna os detalhes de uma recompensa
      tags: *ref_5
    patch:
      operationId: AdminRewardsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRewardDto'
      responses:
        '200':
          description: Recompensa atualizada
        '404':
          description: Recompensa não encontrada
      security: *ref_4
      summary: Atualiza uma recompensa existente
      tags: *ref_5
    delete:
      operationId: AdminRewardsController_softDelete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Recompensa removida
      security: *ref_4
      summary: Soft delete de uma recompensa
      tags: *ref_5
...
  /rewards/{id}:
    get:
      operationId: RewardsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados da recompensa
        '404':
          description: Recompensa não encontrada ou indisponível
      security: *ref_6
      summary: Retorna os detalhes de uma recompensa disponível
      tags: *ref_7
...
  /admin/redemptions:
    get:
      operationId: AdminRedemptionsController_findAll
      parameters:
        - name: status
          required: false
          in: query
          schema:
            enum:
              - pending
              - approved
              - rejected
            type: string
        - name: userId
          required: false
          in: query
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Lista paginada de resgates
      security: &ref_10
        - bearer: []
      summary: Lista todos os resgates com filtros opcionais de status e usuário
      tags: &ref_11
        - admin/redemptions
  /admin/redemptions/{id}:
    get:
      operationId: AdminRedemptionsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados do resgate
        '404':
          description: Resgate não encontrado
      security: *ref_10
      summary: Retorna os detalhes de um resgate específico
      tags: *ref_11
  /admin/redemptions/{id}/approve:
    patch:
      operationId: AdminRedemptionsController_approve
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Resgate aprovado, estoque incrementado
      security: *ref_10
      summary: Aprova um resgate pendente
      tags: *ref_11
  /admin/redemptions/{id}/reject:
    patch:
      operationId: AdminRedemptionsController_reject
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRedemptionDto'
      responses:
        '200':
          description: Resgate rejeitado, pontos reembolsados
      security: *ref_10
      summary: Rejeita um resgate pendente e reembolsa os pontos ao membro
      tags: *ref_11
...
  /redemptions/{id}:
    get:
      operationId: RedemptionsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados do resgate
        '404':
          description: Resgate não encontrado
      security: *ref_12
      summary: Retorna os detalhes de um resgate específico do membro
      tags: *ref_13
...
info:
  title: domy-api
  description: API do app doméstico gamificado
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 10
      required:
        - refreshToken
    AuthUserDto:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
            enum:
              - member
              - admin
          description: Roles do usuário
        level:
          enum:
            - Bronze
            - Prata
            - Ouro
            - Diamante
          type: string
          description: Nível atual do membro
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        pointsBalance:
          type: number
        currentStreak:
          type: number
        longestStreak:
          type: number
        isActive:
          type: boolean
      required:
        - roles
        - level
        - id
        - email
        - name
        - avatarUrl
        - pointsBalance
        - currentStreak
        - longestStreak
        - isActive
    AuthResponseDto:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/AuthUserDto'
        accessToken:
          type: string
        refreshToken:
          type: string
      required:
        - user
        - accessToken
        - refreshToken
...
    CreateRewardDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
          format: uri
        pointsCost:
          type: number
          minimum: 1
        stockLimit:
          type: number
          nullable: true
          minimum: 0
        cooldownDays:
          type: number
          nullable: true
          minimum: 0
        isActive:
          type: boolean
      required:
        - title
        - pointsCost
    UpdateRewardDto:
      type: object
      properties: {}
    RejectRedemptionDto:
      type: object
      properties:
        rejectionReason:
          type: string
          description: Motivo da rejeição do resgate
          example: Recompensa indisponível no momento
      required:
        - rejectionReason
...